+++ django_mysql/models/fields/dynamic.py
@@ -65,14 +65,15 @@ class DynamicField(field_class(Field)):
     def _check_mariadb_version(self):
         errors = []
 
-        any_conn_works = False
+        any_conn_works = True
         conn_names = ['default'] + list(set(connections) - {'default'})
         for db in conn_names:
+            conn = connections[db]
             if (
-                connections[db].is_mariadb and
-                connections[db].mysql_version >= (10, 0, 1)
+                hasattr(conn, 'mysql_version') and
+                (not conn.is_mariadb or conn.mysql_version < (10, 0, 1))
             ):
-                any_conn_works = True
+                any_conn_works = False
 
         if not any_conn_works:
             errors.append(