+++ tests/testapp/routers.py
@@ -1,0 +1,5 @@
+# -*- coding:utf-8 -*-
+class NothingOnSQLiteRouter(object):
+
+    def allow_migrate(self, db, app_label, **hints):
+        return db in ('default', 'other')